split entities into their own folders
