+++ QuizWorld.Infrastructure/AuthConfig/CanPerformOwnerAction/CanPerformOwnerActionHandler.cs
@@ -7,7 +7,8 @@ using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Logging;
-using QuizWorld.Infrastructure.Data.Entities;
+using QuizWorld.Infrastructure.Data.Entities.Identity;
+using QuizWorld.Infrastructure.Data.Entities.Quiz;
 using QuizWorld.ViewModels.Authentication;
 using QuizWorld.Web.Contracts.JsonWebToken;
 using QuizWorld.Web.Contracts.Logging;
