+++ QuizWorld.Web/Program.cs
@@ -4,7 +4,6 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using QuizWorld.Infrastructure.Data;
-using QuizWorld.Infrastructure.Data.Entities;
 using System.Text;
 using System.Text.Json;
 using Newtonsoft.Json.Converters;
@@ -28,6 +27,7 @@ using QuizWorld.Web.Services.GradeService;
 using QuizWorld.Infrastructure.AuthConfig.CanAccessLogs;
 using QuizWorld.Web.Contracts.Logging;
 using QuizWorld.Web.Services.Logging;
+using QuizWorld.Infrastructure.Data.Entities.Identity;
 
 namespace QuizWorld.Web
 {
